requirements.txt
joblib==0.17.0
Keras==2.4.3
matplotlib==3.4.3
numpy==1.18.5
pandas==1.1.3
Pillow==8.4.0
scikit_learn==1.0.1
seaborn==0.11.0
streamlit==1.2.0
tensorflow==2.3.1
tensorflow_hub==0.12.0
tqdm==4.50.2
